Refactor: Split VoiceBot component into smaller components
This commit refactors the VoiceBot component into smaller, more manageable components for improved code organization and maintainability. The changes include separating the messaging UI, audio processing logic, and the main container into their own files.
